Create TwoPointerApproach
This program finds the couple of integers from an array which forms the number as sum given by input. It takes less time complexity than conventional method of using 2 loops for solving these kind of problems.
Explanation with example:

input=16
array={1, 3, 5, 7, 10, 11, 12, 13}
two variables i and j takes the integers from both sides of the array and test if the current combination of numbers is equal to the input.
here 2 combinations, i.e., (1 and 11) and (3 and 13) forms the sum as 16.